check map ok

+++ includes/so_long.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   so_long.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pfrances <pfrances@student.42.fr>          +#+  +:+       +#+        */
+/*   By: pfrances <pfrances@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 16:25:31 by pfrances          #+#    #+#             */
-/*   Updated: 2022/12/01 17:00:33 by pfrances         ###   ########.fr       */
+/*   Updated: 2022/12/01 22:50:56 by pfrances         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,8 +52,8 @@ typedef struct s_borders
 
 typedef struct s_position
 {
-	int	x;
-	int	y;
+	size_t	x;
+	size_t	y;
 }	t_position;
 
 typedef struct s_img
